[#2] feature : 아이템 Wishlist 관련 기능

+++ src/main/java/com/sparta/ecommerce/domain/stock/dto/StockDto.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 @Setter
 public class StockDto {
 
-    private int itemId;
+    private long itemId;
     private int stockQuantity;
 }
