+++ src/main/java/com/sparta/ecommerce/domain/user/dto/SignUpResponseDto.java
@@ -6,14 +6,14 @@ import lombok.Setter;
 @Getter
 @Setter
 public class SignUpResponseDto {
-    private int id;
+    private long id;
     private String userName;
     private String phone;
     private String address;
     private String email;
 
 
-    public SignUpResponseDto(int id, String email, String userName, String phone, String address) {
+    public SignUpResponseDto(long id, String email, String userName, String phone, String address) {
         this.id=id;
         this.userName = userName;
         this.phone = phone;
